+++ Editor/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 26a783b62319a1948b78ae082271220b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
