+++ Editor/UTJ.UnityProfilerLiteKun.Editor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b3fce456dcd56714fa3d5cfb38244565
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
