+++ Runtime/UTJ.UnityProfilerLiteKun.asmdef
@@ -1,0 +1,3 @@
+﻿{
+	"name": "UTJ.UnityProfilerLiteKun"
+}
