+++ Runtime/UTJ.UnityProfilerLiteKun.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0eb4ba82cfe31bd418c8f53b61e612f6
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
